test(sqlite): isolate request log admission coverage
