+++ src/tests/mod.rs
@@ -38,6 +38,7 @@ mod observability_and_lifecycle;
 mod proxy_affinity_and_summary;
 mod proxy_affinity_runtime_geo;
 mod request_kind_and_core;
+mod request_logs_gc_admission;
 mod request_rollup;
 mod request_rollup_public_metrics;
 mod schema_migrations;
